refactor: organiza pagina de perfil dentro de components > paginas.
